feat: Link not exist
Update checking file and add a css class if file not exist in the folder (heading only)

+++ YAFPA/common/file_checking.py
@@ -60,16 +60,32 @@ def delete_not_exist():
     for file in glob.iglob(f"{BASEDIR}/_*/**"):
         if not (any(i in file for i in important_folder)):
             #Delete the file from the database if moved in the excluded folder
-            if os.path.basename(file) not in vault_file or os.path.basename(file) in excluded: #or if file in file_excluded
+            if os.path.basename(file) != '404.md' and (os.path.basename(file) not in vault_file or os.path.basename(file) in excluded): #or if file in file_excluded
                 os.remove(file)
                 info.append(os.path.basename(file))
     return info
 
-def check_file(filepath, folder):
-    post_file = []
-    folder = Path(folder)
-    for file in glob.iglob(f"{folder}/**"):
-        post_file.append(os.path.basename(file))
+def all_file():
+    all_file = {}
+    post = []
+    folder = "_notes"
+    important_folder = ["_includes", "_layout", "_site"]
+    contents = glob.glob(f'{BASEDIR}/_*/**')
+    contents.extend(glob.glob(f'{BASEDIR}/_*/.*'))
+    contents = sorted(contents)
+    for file in contents:
+        if not (any(i in file for i in important_folder)):
+            if folder == os.path.basename(Path(file).parent.absolute()):
+                post.append(file)
+            else:
+                post=[file]
+                folder = os.path.basename(Path(file).parent.absolute())
+            all_file[folder] = post
+    return all_file
+
+def check_file(filepath, folder, all_file):
+    post_file = all_file[os.path.basename(Path(folder))]
+    post_file = [os.path.basename(i) for i in post_file]
     if filepath in post_file:
         return "EXIST"
     else:
